Moved DateUtils to dateHelpers
